Правит ошибки в реализации полноразмерных фото
