+++ js/const/const.js
@@ -1,6 +1,6 @@
 const COUNT_PUBLISH_PHOTO = 25;
 const MAX_COUNT_ID = 25;
-const COMMENTS_MAX_COUNT = 5;
+const COMMENTS_MAX_COUNT = 15;
 const LIKES_MIN_COUNT = 15;
 const LIKES_MAX_COUNT = 200;
 const AVATARS_MAX_COUNT = 6;
