+++ js/utils/helpers.js
@@ -22,6 +22,26 @@ function getUniqueId (min, max) {
   return uniqueId;
 }
 
+// функция нахождения случайного уникального числа 2.0
+// function createRandomIdFromRangeGenerator (min, max) {
+
+//   return function () {
+//     const previousValues = [];
+
+//     let currentValue = getRandomInteger(min, max);
+//     if (previousValues.length >= (max - min + 1)) {
+//       return null;
+//     }
+//     while (previousValues.includes(currentValue)) {
+//       currentValue = getRandomInteger(min, max);
+//     }
+//     previousValues.push(currentValue);
+//     return currentValue;
+//   };
+// }
+
+// const getUniqueId = createRandomIdFromRangeGenerator(1, 25);
+
 // функция нахождения рандомного элемента массива
 const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
